[Refactor] Move tasks back to static object
